requirements.txt
boto3==1.28.29
cryptography==41.0.3
fastapi==0.101.1
logtail-python==0.2.6
newrelic~=9.5.0
newrelic
passlib~=1.7.4
passlib[bcrypt]
pika~=1.3.2
pydantic~=2.3.0
pydantic-settings==2.0.3
PyJWT==2.8.0
python-decouple==3.8
python-jose
python-multipart
redis==5.0.0
requests==2.31.0
sentry-sdk[fastapi]
starlette~=0.27.0
Unidecode~=1.3.6
uvicorn==0.23.2
validate-docbr
